buns_and_cakes/requirements.txt
certifi==2020.6.20
chardet==4.0.0
idna==2.10
peewee==3.14.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dotenv==0.17.1
requests==2.25.1
six==1.15.0
transitions==0.8.8
urllib3==1.26.4
vk-api==11.9.3
